+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### 4.0.0 (pending)
 
 * renamed to Maven Summon
+* removed groupName argument from deploy methods, because it uses compat code
 * Maven is AutoClosable now, close() invokes container.dispose() to fix memory leak
 * added PluginRepositoryBlocker
 * added ExtensionBlocker that can block extension classes/resources being loaded; default is to block all